+++ TRBot/TRBot.VirtualControllers/Controllers/VJoyController.cs
@@ -235,7 +235,7 @@ namespace TRBot.VirtualControllers
             InputReleasedEvent?.Invoke(inputName);
         }
 
-        public void PressAxis(in int axis, in double minAxisVal, in double maxAxisVal, in int percent)
+        public void PressAxis(in int axis, in double minAxisVal, in double maxAxisVal, in double percent)
         {
             //Not a valid axis - defaulting to 0 results in the wrong axis being set
             if (AxisCodeMap.TryGetValue(axis, out int vJoyAxis) == false)
@@ -387,7 +387,7 @@ namespace TRBot.VirtualControllers
             return InputTracker.GetButtonState(buttonVal);
         }
 
-        public int GetAxisState(in int axisVal)
+        public double GetAxisState(in int axisVal)
         {
             return InputTracker.GetAxisState(axisVal);
         }
