+++ TRBot/TRBot.VirtualControllers/VirtualControllerDelegates.cs
@@ -30,7 +30,7 @@ namespace TRBot.VirtualControllers
         public delegate void OnInputPressed(in string inputName);
         public delegate void OnInputReleased(in string inputName);
 
-        public delegate void OnAxisPressed(in int axis, in int percent);
+        public delegate void OnAxisPressed(in int axis, in double percent);
         public delegate void OnAxisReleased(in int axis);
 
         public delegate void OnButtonPressed(in uint buttonVal);
